Extract execute_transaction() from the bank
